refactor(aug): rename `aug_config` to `aug_configs` across project
