+++ tests/datasets/test_kornia_transforms.py
@@ -12,7 +12,7 @@ All tests in this module are CPU-compatible — Kornia operates on CPU tensors i
 import pytest
 import torch
 
-from rfdetr.datasets.aug_config import (
+from rfdetr.datasets.aug_configs import (
     AUG_AERIAL,
     AUG_AGGRESSIVE,
     AUG_CONSERVATIVE,
